Fiz alterações na página de configurações
