--- AcquaSense React/src/pages/Configu/Configu.css
@@ -1,126 +1,0 @@
-
-/* Variáveis de tema claro */
-:root {
-    --bg-color: #f8f8f8;
-    --text-color: #333;
-    --sidebar-bg-color: #f3eaff;
-    --logo-color: #6c63ff;
-    --nav-link-color: #333;
-    --nav-link-hover-bg-color: #e0d4ff;
-    --nav-link-hover-color: #6c63ff;
-    --header-bg-color: #fff;
-    --header-border-color: #ddd;
-    --card-bg-color: #fff;
-    --info-card-bg-color: #eef6ff;
-    --info-button-bg-color: #6c63ff;
-    --info-button-hover-bg-color: #5752d4;
-  }
-  
-  /* Variáveis de tema escuro */
-  body.dark-mode {
-    --bg-color: #181818;
-    --text-color: #f0f0f0;
-    --sidebar-bg-color: #222222;
-    --logo-color: #bb86fc;
-    --nav-link-color: #f0f0f0;
-    --nav-link-hover-bg-color: #333333;
-    --nav-link-hover-color: #bb86fc;
-    --header-bg-color: #222222;
-    --header-border-color: #333333;
-    --card-bg-color: #333333;
-    --info-card-bg-color: #2c2c2c;
-    --info-button-bg-color: #bb86fc;
-    --info-button-hover-bg-color: #8e7af7;
-  }
-  
-  /* Configurações de Notificações */
-  .notification-settings,
-  .data-analysis-settings {
-      background-color: var(--card-bg-color); /* Cor de fundo usando variável */
-      border-radius: 8px;
-      box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-      padding: 20px;
-      margin-right: 20px; /* Espaçamento entre os blocos */
-      flex: 1; /* Permite que o bloco cresça e se ajuste */
-      min-height: 400px; /* Define uma altura mínima para os contêineres */
-  }
-  
-  .notification-settings h2,
-  .data-analysis-settings h2 {
-      margin-top: 0;
-      margin-bottom: 20px; /* Aumenta o espaçamento entre o título e os subtítulos */
-      color: var(--text-color); /* Cor do texto usando variável */
-  }
-  
-  .notification-settings h3,
-  .data-analysis-settings h3 {
-      margin-top: 10px; /* Aumenta o espaçamento acima dos subtítulos */
-      margin-bottom: 10px; /* Espaço abaixo dos subtítulos */
-      font-size: 18px; /* Tamanho dos subtítulos */
-      color: var(--text-color); /* Cor do texto usando variável */
-  }
-  
-  .notification-settings label,
-  .data-analysis-settings label {
-      display: block;
-      margin-bottom: 10px;
-      color: var(--text-color); /* Cor do texto usando variável */
-  }
-  
-  /* Estilos dos botões */
-  .buttons {
-      display: flex;
-      flex-direction: row; /* Alinha os botões horizontalmente */
-      margin-top: 20px;
-  }
-  
-  .buttons button {
-      margin-right: 10px;
-      padding: 10px 20px;
-      border: none;
-      border-radius: 4px;
-      cursor: pointer;
-      background-color: var(--info-button-bg-color); /* Cor do botão usando variável */
-      color: #fff; /* Cor do texto do botão */
-      font-size: 16px;
-  }
-  
-  .buttons button:hover {
-      background-color: var(--info-button-hover-bg-color); /* Cor do botão ao passar o mouse usando variável */
-  }
-  
-  /* Ajustes no contêiner principal */
-  .config-container {
-      display: flex;
-      flex-direction: row;
-      align-items: stretch; /* Faz os contêineres filhos terem a mesma altura */
-      justify-content: flex-start;
-      padding-top: 20px; /* Ajuste este valor se necessário */
-      padding-left: 350px;
-      padding-right: 20px;
-      margin-top: -725px; /* Ajuste para trazer o conteúdo mais para cima */
-  }
-  
-  /* Estilos para o botão de editar configurações */
-  .edit-button-container {
-      display: flex; /* Habilita o Flexbox */
-      justify-content: center; /* Centraliza horizontalmente */
-      margin-top: 20px; /* Espaçamento acima do contêiner do botão */
-      z-index: 10; /* Para garantir que o botão esteja acima de outros elementos */
-  }
-  
-  .edit-button {
-      padding: 10px 20px;
-      background-color: var(--info-button-bg-color); /* Cor do botão usando variável */
-      color: #fff; /* Cor do texto do botão */
-      border: none;
-      border-radius: 4px;
-      cursor: pointer;
-      font-size: 16px;
-  }
-  
-  .edit-button:hover {
-      background-color: var(--info-button-hover-bg-color); /* Cor do botão ao passar o mouse usando variável */
-  }
-  
-
